pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skillforge"
version = "0.1.0"
description = "Agent Skill 增强系统 — 让 Agent 先自我诊断，再决定是否启用 skill 增强"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "SkillForge Team", email = "skillforge@example.com" }
]
keywords = ["agent", "skill", "llm", "self-improvement", "capability-awareness"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "typer[all]>=0.12",
    "rich>=13.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
skillforge = "skillforge.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/skillforge"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
    "/config.yaml",
    "/skillforge-registry.yaml",
    "/memory",
]

[tool.hatch.editable]
packages = ["src/skillforge"]
editable-paths = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
